# ss/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "movies-app-ss",
        "version": "1.0.0"
    },
    "paths": {
...
        "api/password_resets": {
            "get": {
                "tags": [
                    "PasswordReset"
                ],
                "summary": "Get list of password resets.",
                "operationId": "dd36f74550eae37544dbe9916f6fc1d2",
                "responses": {
                    "200": {
                        "description": "Display a listing of password_resets."
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "post": {
                "tags": [
                    "PasswordReset"
                ],
                "summary": "Create new password_reset.",
                "operationId": "a4a3482162578095bc8e14d55379e17f",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email of password_reset",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/PasswordReset/properties/email"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Status of category.",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/PasswordReset/properties/token"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PasswordReset"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create successfully."
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "api/password_resets/{id}": {
            "get": {
                "tags": [
                    "PasswordReset"
                ],
                "summary": "Get password_reset by id",
                "operationId": "3b68e422bfad7096b14d71c128df6d2c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/PasswordReset/properties/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Display a listing of films."
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "put": {
                "tags": [
                    "PasswordReset"
                ],
                "summary": "Update password_reset by id.",
                "operationId": "706adcf9a4be241b27383f07a00551c6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/PasswordReset/properties/id"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email of password_reset",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/PasswordReset/properties/email"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Status of password_reset.",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/PasswordReset/properties/token"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PasswordReset"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update successfully."
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "PasswordReset"
                ],
                "summary": "Delete password_reset by id.",
                "operationId": "b99d63e48567de49b1e82a09b18a0914",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/PasswordReset/properties/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete successfully."
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "api/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get list of users.",
                "operationId": "7a4026e2942db08b9bb6dbecaa330a7f",
                "responses": {
                    "200": {
                        "description": "Display a listing of users."
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create new user.",
                "operationId": "dc05e3f0479b55da3841174632933631",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of user",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/User/properties/name"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email of category.",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/User/properties/email"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create successfully."
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "api/users/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get user by id",
                "operationId": "2d9515c3b86da14a85b45e5fce976985",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/User/properties/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Display a listing of films."
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update user by id.",
                "operationId": "d80d10b99c4758003976f4ea7a405966",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/User/properties/id"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of user",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/User/properties/name"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email of user.",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/User/properties/email"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update successfully."
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete user by id.",
                "operationId": "12f186e0f7bfd882720e227da3937bfa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/User/properties/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete successfully."
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "PasswordReset": {
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "$ref": "#/components/schemas/BaseModel/properties/id"
                    },
                    "email": {
                        "type": "string"
                    },
                    "token": {
                        "type": "string"
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_at"
                    },
                    "update_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PasswordReset"
                }
            },
...
            "User": {
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "id": {
                        "$ref": "#/components/schemas/BaseModel/properties/id"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "type": "string"
                    },
                    "status": {
                        "type": "integer",
                        "example": "Trạng thái: 1-kích hoạt, 2-vô hiệu, 3-đã xóa"
                    },
                    "position": {
                        "type": "integer",
                        "example": "1"
                    },
                    "created_by": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_by"
                    },
                    "updated_by": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    },
                    "deleted_by": {
                        "$ref": "#/components/schemas/BaseModel/properties/deleted_at"
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_at"
                    },
                    "update_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    },
                    "deleted_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/deleted_at"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            }
        }
    }
}